+++ src/main/java/org/gp/civiceye/repository/MasterAdminRepository.java
@@ -3,5 +3,5 @@ package org.gp.civiceye.repository;
 import org.gp.civiceye.repository.entity.MasterAdmin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MasterAdminRepository extends JpaRepository<MasterAdmin, Integer> {
+public interface MasterAdminRepository extends JpaRepository<MasterAdmin, Long> {
 }
