+++ src/main/java/org/gp/civiceye/repository/StatusHistoryRepository.java
@@ -3,5 +3,5 @@ package org.gp.civiceye.repository;
 import org.gp.civiceye.repository.entity.StatusHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StatusHistoryRepository extends JpaRepository<StatusHistory, Integer> {
+public interface StatusHistoryRepository extends JpaRepository<StatusHistory, Long> {
 }
